% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{atusyverse_deps}
\alias{atusyverse_deps}
\title{List all atusyverse dependencies}
\usage{
atusyverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
atusyverse packages.}
}
\description{
List all atusyverse dependencies
}
