% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{atusyverse_packages}
\alias{atusyverse_packages}
\title{List all packages in the atusyverse}
\usage{
atusyverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include atusyverse in the list?}
}
\description{
List all packages in the atusyverse
}
\examples{
atusyverse_packages()
}
